refactor product.dto
