+++ src/products/dto/generic-product.dto.ts
@@ -9,7 +9,7 @@ import {
 } from 'class-validator';
 import { BaseDTO } from 'src/utils/dto/base.dto';
 import { ResponseManufacturerDTO } from './manufacturer.dto';
-import { CategoryDTO } from './find-products.dto';
+import { CategoryResponseDTO } from 'src/category/dto/category.dto';
 
 export class GenericProductDTO {
   @IsString()
@@ -55,7 +55,7 @@ export class ResponseGenericProductDTO extends IntersectionType(
 
   @ApiProperty({
     description: 'List of categories of the product',
-    type: [CategoryDTO],
+    type: [CategoryResponseDTO],
   })
-  categories: CategoryDTO[];
+  categories: CategoryResponseDTO[];
 }
